lxd/devlxd: Regroup vSock and devLXD functions
